+++ examples/javaScript/application/oauthToken.js
@@ -5,17 +5,18 @@ var clientSecret = process.env.EBAY_CLIENT_SECRET || 'YOUR_SECRET';
 
 var eBay = require('../../../lib/eBay-node-client')(clientId, clientSecret);
 
-var feedRequest = async function () {
+var applicationRequest = async function () {
 
     try {
         var token = await eBay.application.getOAuthToken({
             grant_type: 'client_credentials',
             scope: 'https://api.ebay.com/oauth/api_scope'
         });
+        console.log('token.access_token ', token.access_token);
         eBay.setToken(token.access_token);
     } catch (error) {
         console.log('error ', error);
     }
 };
 
-feedRequest();
+applicationRequest();