+++ examples/javaScript/buy/browse/search.js
@@ -5,7 +5,7 @@ var clientSecret = process.env.EBAY_CLIENT_SECRET || 'YOUR_SECRET';
 
 var eBay = require('../../../../lib/eBay-node-client')(clientId, clientSecret);
 
-var categoryRequest = async function () {
+var browseRequest = async function () {
     try {
         var token = await eBay.application.getOAuthToken({
             grant_type: 'client_credentials',
@@ -17,9 +17,10 @@ var categoryRequest = async function () {
     }
 
     var data = {
-        marketplace_id: 'EBAY_IN'
+        gtin: '010942122258',
+        limit: '1'
     };
-    eBay.taxonomy.getDefaultCategoryTreeId(data, function (error, response) {
+    eBay.browse.search(data, function (error, response) {
         if (error) {
             console.log('error ', error);
             return;
@@ -28,4 +29,4 @@ var categoryRequest = async function () {
     });
 };
 
-categoryRequest();
+browseRequest();