+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.0.2 September 19, 2018
+  - Add Support for browse API : search
+  - Add SANDBOX usage using environment variable
+  - Changes for the Taxonomy API
+  - Add Support for Inventory API : createOrReplaceInventoryItem, getInventoryItem
+
 0.0.1 September 19, 2018
   - Initial release with Taxonomy API
   - Add Support for getOAuthToken